requirements.txt
openai
pymilvus
streamlit-analytics2
google-cloud-firestore
